updated product views security - super users only

+++ products/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, redirect, reverse, get_object_or_404
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.db.models.functions import Lower
 
@@ -10,12 +11,13 @@ from .forms import ProductForm
 
 
 def all_products(request):
-    """ view to show all the products including sorting and search queries """
+    """ View to show all the products including sorting and search queries """
 
     products = Product.objects.all()
     query = None
     categories = None
 
+    """ This is for the search bar to find products """
     if request.GET:
         if 'category' in request.GET:
             categories = request.GET['category'].split(',')
@@ -52,7 +54,13 @@ def product_detail(request, product_id):
     return render(request, 'products/product_detail.html', context)
 
 
+@login_required
 def add_product(request):
+    """ Only logged in super users can access this feature """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners can do that.')
+        return redirect(reverse('home'))
+
     """ Add a product to the store """
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
@@ -73,7 +81,13 @@ def add_product(request):
     return render(request, template, context)
 
 
+@login_required
 def edit_product(request, product_id):
+    """ Only logged in super users can access this feature """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners can do that.')
+        return redirect(reverse('home'))
+
     """ Edit a product in the store """
     product = get_object_or_404(Product, pk=product_id)
     if request.method == 'POST':
@@ -97,7 +111,13 @@ def edit_product(request, product_id):
     return render(request, template, context)
 
 
+@login_required
 def delete_product(request, product_id):
+    """ Only logged in super users can access this feature """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners can do that.')
+        return redirect(reverse('home'))
+
     """ Delete a product from the store """
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
